requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
google-api-python-client==2.86.0
google-auth-httplib2==0.2.0
google-cloud-storage==2.9.0
google-generativeai==0.8.5
beautifulsoup4==4.12.2
gunicorn==20.1.0
waitress==3.0.0
python-dotenv==1.0.0
openpyxl==3.1.2
python-docx==1.1.0
pypdf==4.0.1
PyPDF2==3.0.1
pdfplumber==0.10.3
reportlab==4.0.7